third_party/WebKit/LayoutTests/external/wpt/audio-output/setSinkId-expected.txt

This is a testharness.js-based test.
PASS setSinkId on default audio output should always work
PASS setSinkId fails with NotFoundError on made up deviceid
FAIL List device, setSinkId should be allowed on the default, the rest of the devices will get a NotAllowedError assert_equals: Non-default devices are failing with NotAllowed error expected "NotAllowedError" but got "NotFoundError"
Harness: the test ran to completion.
